# fix: memory allocations