Benchmarks 
```bash
go test -v -bench=. -benchmem -cpu 1
```

before
```bash
goos: darwin
goarch: amd64
pkg: github.com/go-ldap/ldap
BenchmarkFilterCompile            338554              3420 ns/op            1606 B/op         61 allocs/op
BenchmarkFilterDecompile         1578404               775 ns/op             155 B/op          8 allocs/op

BenchmarkFilterCompile            335392              3518 ns/op            1606 B/op         61 allocs/op
BenchmarkFilterDecompile         1557634               765 ns/op             155 B/op          8 allocs/op

BenchmarkFilterCompile            342820              3423 ns/op            1606 B/op         61 allocs/op
BenchmarkFilterDecompile         1535986               769 ns/op             155 B/op          8 allocs/op
```

after
```bash
goos: darwin
goarch: amd64
pkg: github.com/go-ldap/ldap
BenchmarkFilterCompile            622689              1914 ns/op            1581 B/op         36 allocs/op
BenchmarkFilterDecompile         1779561               671 ns/op             171 B/op          3 allocs/op

BenchmarkFilterCompile            616844              1932 ns/op            1581 B/op         36 allocs/op
BenchmarkFilterDecompile         1753670               683 ns/op             171 B/op          3 allocs/op

BenchmarkFilterCompile            635676              1927 ns/op            1581 B/op         36 allocs/op
BenchmarkFilterDecompile         1793448               676 ns/op             171 B/op          3 allocs/op
```