@johnweldon 2019-12-13
> Copy all updated files into `v3` folder to respect Go modules.

Thanks for catching this.